refactor: convert entity IDs and shuffle entity+renderer code around
