+++ src/main/java/com/chimericdream/minekea/entities/blocks/containers/CrateBlockEntity.java
@@ -1,5 +1,8 @@
-package com.chimericdream.minekea.block.containers.crates;
+package com.chimericdream.minekea.entities.blocks.containers;
 
+import com.chimericdream.minekea.ModInfo;
+import com.chimericdream.minekea.block.containers.crates.Crates;
+import com.chimericdream.minekea.block.containers.crates.GenericCrate;
 import com.chimericdream.minekea.screen.crate.CrateScreenHandler;
 import com.chimericdream.minekea.util.ImplementedInventory;
 import net.minecraft.block.BlockState;
@@ -19,6 +22,7 @@ import net.minecraft.sound.SoundEvent;
 import net.minecraft.sound.SoundEvents;
 import net.minecraft.text.Text;
 import net.minecraft.text.TranslatableText;
+import net.minecraft.util.Identifier;
 import net.minecraft.util.collection.DefaultedList;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.Direction;
@@ -26,6 +30,8 @@ import net.minecraft.util.math.Vec3i;
 import net.minecraft.world.World;
 
 public class CrateBlockEntity extends BlockEntity implements NamedScreenHandlerFactory, ImplementedInventory {
+    public static final Identifier ENTITY_ID = new Identifier(ModInfo.MOD_ID, "entities/blocks/containers/crate");
+
     private final DefaultedList<ItemStack> items = DefaultedList.ofSize(GenericCrate.ROW_COUNT * 9, ItemStack.EMPTY);
     private final ViewerCountManager stateManager;
