+++ src/main/java/com/chimericdream/minekea/entities/blocks/furniture/DisplayCaseBlockEntity.java
@@ -1,7 +1,10 @@
-package com.chimericdream.minekea.block.furniture.shelves;
+package com.chimericdream.minekea.entities.blocks.furniture;
 
+import com.chimericdream.minekea.ModInfo;
+import com.chimericdream.minekea.block.furniture.displaycases.DisplayCases;
 import com.chimericdream.minekea.util.ImplementedInventory;
 import net.minecraft.block.BlockState;
+import net.minecraft.block.Blocks;
 import net.minecraft.block.entity.BlockEntity;
 import net.minecraft.block.entity.BlockEntityType;
 import net.minecraft.entity.player.PlayerEntity;
@@ -16,6 +19,7 @@ import net.minecraft.server.world.ServerWorld;
 import net.minecraft.sound.SoundCategory;
 import net.minecraft.sound.SoundEvent;
 import net.minecraft.sound.SoundEvents;
+import net.minecraft.util.Identifier;
 import net.minecraft.util.collection.DefaultedList;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.Direction;
@@ -23,14 +27,16 @@ import net.minecraft.world.World;
 
 import javax.annotation.Nullable;
 
-public class ShelfBlockEntity extends BlockEntity implements ImplementedInventory, SidedInventory {
-    private final DefaultedList<ItemStack> items = DefaultedList.ofSize(4, ItemStack.EMPTY);
+public class DisplayCaseBlockEntity extends BlockEntity implements ImplementedInventory, SidedInventory {
+    public static final Identifier ENTITY_ID = new Identifier(ModInfo.MOD_ID, "entities/blocks/furniture/display_case");
 
-    public ShelfBlockEntity(BlockPos pos, BlockState state) {
-        this(Shelves.SHELF_BLOCK_ENTITY, pos, state);
+    private final DefaultedList<ItemStack> items = DefaultedList.ofSize(1, ItemStack.EMPTY);
+
+    public DisplayCaseBlockEntity(BlockPos pos, BlockState state) {
+        this(DisplayCases.DISPLAY_CASE_BLOCK_ENTITY, pos, state);
     }
 
-    public ShelfBlockEntity(BlockEntityType<? extends ShelfBlockEntity> type, BlockPos pos, BlockState state) {
+    public DisplayCaseBlockEntity(BlockEntityType<? extends DisplayCaseBlockEntity> type, BlockPos pos, BlockState state) {
         super(type, pos, state);
     }
 
@@ -39,67 +45,20 @@ public class ShelfBlockEntity extends BlockEntity implements ImplementedInventor
         return items;
     }
 
-    public NbtCompound getNbt() {
-        NbtCompound nbt = new NbtCompound();
-
-        this.writeNbt(nbt);
-
-        return nbt;
-    }
-
     @Override
     public void readNbt(NbtCompound nbt) {
         items.clear();
+        super.readNbt(nbt);
         Inventories.readNbt(nbt, items);
-        markDirty();
+        if (items.get(0).isOf(Blocks.BARRIER.asItem())) {
+            items.set(0, ItemStack.EMPTY);
+        }
     }
 
     @Override
     public void writeNbt(NbtCompound nbt) {
         Inventories.writeNbt(nbt, items);
-    }
-
-    @Nullable
-    @Override
-    public Packet<ClientPlayPacketListener> toUpdatePacket() {
-        return BlockEntityUpdateS2CPacket.create(this);
-    }
-
-    @Override
-    public NbtCompound toInitialChunkDataNbt() {
-        return createNbt();
-    }
-
-    @Override
-    public int[] getAvailableSlots(Direction var1) {
-        int[] result = new int[getItems().size()];
-
-        for (int i = 0; i < result.length; i++) {
-            result[i] = i;
-        }
-
-        return result;
-    }
-
-    @Override
-    public boolean canInsert(int slot, ItemStack stack, Direction direction) {
-        return false;
-    }
-
-    @Override
-    public boolean canExtract(int slot, ItemStack stack, Direction direction) {
-        return true;
-    }
-
-    @Override
-    public ItemStack tryInsert(int slot, ItemStack stack) {
-        ItemStack ret = ImplementedInventory.super.tryInsert(slot, stack);
-
-        if (!ItemStack.areEqual(ret, stack)) {
-            this.playAddItemSound();
-        }
-
-        return ret;
+        super.writeNbt(nbt);
     }
 
     @Override
@@ -115,6 +74,40 @@ public class ShelfBlockEntity extends BlockEntity implements ImplementedInventor
         if (!world.isClient()) {
             ((ServerWorld) world).getChunkManager().markForUpdate(pos); // Mark changes to be synced to the client.
         }
+    }
+
+    @Nullable
+    @Override
+    public Packet<ClientPlayPacketListener> toUpdatePacket() {
+        return BlockEntityUpdateS2CPacket.create(this);
+    }
+
+    @Override
+    public NbtCompound toInitialChunkDataNbt() {
+        return createNbt();
+    }
+
+    @Override
+    public int[] getAvailableSlots(Direction var1) {
+        return new int[]{};
+    }
+
+    @Override
+    public boolean canInsert(int slot, ItemStack stack, Direction direction) {
+        return false;
+    }
+
+    @Override
+    public boolean canExtract(int slot, ItemStack stack, Direction direction) {
+        return false;
+    }
+
+    public void playRemoveItemSound() {
+        playSound(SoundEvents.ENTITY_ITEM_FRAME_REMOVE_ITEM);
+    }
+
+    public void playRotateItemSound() {
+        playSound(SoundEvents.ENTITY_ITEM_FRAME_ROTATE_ITEM);
     }
 
     public void playAddItemSound() {
