+++ src/main/java/com/chimericdream/minekea/entities/blocks/furniture/ShelfBlockEntity.java
@@ -1,8 +1,9 @@
-package com.chimericdream.minekea.block.furniture.displaycases;
+package com.chimericdream.minekea.entities.blocks.furniture;
 
+import com.chimericdream.minekea.ModInfo;
+import com.chimericdream.minekea.block.furniture.shelves.Shelves;
 import com.chimericdream.minekea.util.ImplementedInventory;
 import net.minecraft.block.BlockState;
-import net.minecraft.block.Blocks;
 import net.minecraft.block.entity.BlockEntity;
 import net.minecraft.block.entity.BlockEntityType;
 import net.minecraft.entity.player.PlayerEntity;
@@ -17,6 +18,7 @@ import net.minecraft.server.world.ServerWorld;
 import net.minecraft.sound.SoundCategory;
 import net.minecraft.sound.SoundEvent;
 import net.minecraft.sound.SoundEvents;
+import net.minecraft.util.Identifier;
 import net.minecraft.util.collection.DefaultedList;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.Direction;
@@ -24,14 +26,16 @@ import net.minecraft.world.World;
 
 import javax.annotation.Nullable;
 
-public class DisplayCaseBlockEntity extends BlockEntity implements ImplementedInventory, SidedInventory {
-    private final DefaultedList<ItemStack> items = DefaultedList.ofSize(1, ItemStack.EMPTY);
+public class ShelfBlockEntity extends BlockEntity implements ImplementedInventory, SidedInventory {
+    public static final Identifier ENTITY_ID = new Identifier(ModInfo.MOD_ID, "entities/blocks/furniture/shelf");
 
-    public DisplayCaseBlockEntity(BlockPos pos, BlockState state) {
-        this(DisplayCases.DISPLAY_CASE_BLOCK_ENTITY, pos, state);
+    private final DefaultedList<ItemStack> items = DefaultedList.ofSize(4, ItemStack.EMPTY);
+
+    public ShelfBlockEntity(BlockPos pos, BlockState state) {
+        this(Shelves.SHELF_BLOCK_ENTITY, pos, state);
     }
 
-    public DisplayCaseBlockEntity(BlockEntityType<? extends DisplayCaseBlockEntity> type, BlockPos pos, BlockState state) {
+    public ShelfBlockEntity(BlockEntityType<? extends ShelfBlockEntity> type, BlockPos pos, BlockState state) {
         super(type, pos, state);
     }
 
@@ -40,35 +44,24 @@ public class DisplayCaseBlockEntity extends BlockEntity implements ImplementedIn
         return items;
     }
 
+    public NbtCompound getNbt() {
+        NbtCompound nbt = new NbtCompound();
+
+        this.writeNbt(nbt);
+
+        return nbt;
+    }
+
     @Override
     public void readNbt(NbtCompound nbt) {
         items.clear();
-        super.readNbt(nbt);
         Inventories.readNbt(nbt, items);
-        if (items.get(0).isOf(Blocks.BARRIER.asItem())) {
-            items.set(0, ItemStack.EMPTY);
-        }
+        markDirty();
     }
 
     @Override
     public void writeNbt(NbtCompound nbt) {
         Inventories.writeNbt(nbt, items);
-        super.writeNbt(nbt);
-    }
-
-    @Override
-    public void markDirty() {
-        if (this.world != null) {
-            markDirtyInWorld(this.world, this.pos, this.getCachedState());
-        }
-    }
-
-    protected void markDirtyInWorld(World world, BlockPos pos, BlockState state) {
-        world.markDirty(pos);
-
-        if (!world.isClient()) {
-            ((ServerWorld) world).getChunkManager().markForUpdate(pos); // Mark changes to be synced to the client.
-        }
     }
 
     @Nullable
@@ -84,7 +77,13 @@ public class DisplayCaseBlockEntity extends BlockEntity implements ImplementedIn
 
     @Override
     public int[] getAvailableSlots(Direction var1) {
-        return new int[]{};
+        int[] result = new int[getItems().size()];
+
+        for (int i = 0; i < result.length; i++) {
+            result[i] = i;
+        }
+
+        return result;
     }
 
     @Override
@@ -94,15 +93,33 @@ public class DisplayCaseBlockEntity extends BlockEntity implements ImplementedIn
 
     @Override
     public boolean canExtract(int slot, ItemStack stack, Direction direction) {
-        return false;
+        return true;
     }
 
-    public void playRemoveItemSound() {
-        playSound(SoundEvents.ENTITY_ITEM_FRAME_REMOVE_ITEM);
+    @Override
+    public ItemStack tryInsert(int slot, ItemStack stack) {
+        ItemStack ret = ImplementedInventory.super.tryInsert(slot, stack);
+
+        if (!ItemStack.areEqual(ret, stack)) {
+            this.playAddItemSound();
+        }
+
+        return ret;
     }
 
-    public void playRotateItemSound() {
-        playSound(SoundEvents.ENTITY_ITEM_FRAME_ROTATE_ITEM);
+    @Override
+    public void markDirty() {
+        if (this.world != null) {
+            markDirtyInWorld(this.world, this.pos, this.getCachedState());
+        }
+    }
+
+    protected void markDirtyInWorld(World world, BlockPos pos, BlockState state) {
+        world.markDirty(pos);
+
+        if (!world.isClient()) {
+            ((ServerWorld) world).getChunkManager().markForUpdate(pos); // Mark changes to be synced to the client.
+        }
     }
 
     public void playAddItemSound() {
