+++ src/main/resources/assets/minekea/blockstates/containers/glass_jar.json
@@ -1,0 +1,7 @@
+{
+    "variants": {
+        "": {
+            "model": "minekea:block/containers/glass_jar"
+        }
+    }
+}
