+++ src/main/resources/assets/minekea/models/block/containers/glass_jar.json
@@ -2,7 +2,6 @@
     "textures": {
         "planks": "minecraft:block/acacia_planks",
         "glass": "minecraft:block/glass",
-        "label": "minekea:block/jars/label",
         "particle": "minecraft:block/glass"
     },
     "display": {
