+++ src/main/resources/data/minekea/tags/blocks/non_wooden_vertical_stairs.json
@@ -25,7 +25,6 @@
         "minekea:building/stairs/mossy_stone_brick_vertical_stairs",
         "minekea:building/stairs/polished_diorite_vertical_stairs",
         "minekea:building/stairs/mossy_cobblestone_vertical_stairs",
-        "minekea:building/stairs/end_stone_brick_vertical_stairs",
         "minekea:building/stairs/stone_vertical_stairs",
         "minekea:building/stairs/smooth_sandstone_vertical_stairs",
         "minekea:building/stairs/smooth_quartz_vertical_stairs",
