+++ src/main/resources/data/minekea/tags/blocks/wooden_vertical_stairs.json
@@ -1,14 +1,22 @@
 {
     "replace": false,
     "values": [
-        "minekea:building/stairs/oak_vertical_stairs",
-        "minekea:building/stairs/spruce_vertical_stairs",
-        "minekea:building/stairs/birch_vertical_stairs",
-        "minekea:building/stairs/jungle_vertical_stairs",
-        "minekea:building/stairs/crimson_vertical_stairs",
-        "minekea:building/stairs/warped_vertical_stairs",
-        "minekea:building/stairs/acacia_vertical_stairs",
-        "minekea:building/stairs/dark_oak_vertical_stairs",
+        "minekea:building/stairs/oak_plank_vertical_stairs",
+        "minekea:building/stairs/spruce_plank_vertical_stairs",
+        "minekea:building/stairs/birch_plank_vertical_stairs",
+        "minekea:building/stairs/jungle_plank_vertical_stairs",
+        "minekea:building/stairs/crimson_plank_vertical_stairs",
+        "minekea:building/stairs/warped_plank_vertical_stairs",
+        "minekea:building/stairs/acacia_plank_vertical_stairs",
+        "minekea:building/stairs/dark_oak_plank_vertical_stairs",
+        "minekea:building/stairs/oak_log_vertical_stairs",
+        "minekea:building/stairs/spruce_log_vertical_stairs",
+        "minekea:building/stairs/birch_log_vertical_stairs",
+        "minekea:building/stairs/jungle_log_vertical_stairs",
+        "minekea:building/stairs/crimson_stem_vertical_stairs",
+        "minekea:building/stairs/warped_stem_vertical_stairs",
+        "minekea:building/stairs/acacia_log_vertical_stairs",
+        "minekea:building/stairs/dark_oak_log_vertical_stairs",
         "minekea:building/stairs/acacia_vertical_bookshelf_stairs",
         "minekea:building/stairs/birch_vertical_bookshelf_stairs",
         "minekea:building/stairs/crimson_vertical_bookshelf_stairs",
